regorust
Function regoBundleSave

pub unsafe extern "C" fn regoBundleSave(
    rego: *mut regoInterpreter,
    dir: *const c_char,
    bundle: *mut regoBundle,
) -> regoEnum

@brief Saves a compiled Rego bundle to the specified directory. @details The directory will be created if it does not exist. The bundle will be saved in the standard Rego bundle format, which consists of a plan.json file, a data.json file, and zero or more .rego files. @param rego The interpreter @param dir The path to the directory where the bundle will be saved. @param bundle The bundle to save. @return REGO_OK if successful, REGO_ERROR otherwise.