Function regorust::regoSetTZDataPath

pub unsafe extern "C" fn regoSetTZDataPath(
    path: *const c_char,
) -> regoEnum

Sets the location where rego-cpp will look for timezone database information.

The timezone database will be interpreted as one obtained from the IANA (https://www.iana.org/time-zones) which has been downloaded and unpacked into at the path provided. If the library was built without manual tzdata support, this function will return an error code.

@param path The path to the timezone database. @return REGO_OK if successful, REGO_ERROR_MANUAL_TZDATA_NOT_SUPPORTED otherwise.