Crate regorust

Structs

Bundle

Representation of a compiled Rego bundle. This can be produced by the Interpreter::build() method as an alternative to immediate evaluation.

Input

The Input interface allows the creation of inputs to a policy without requiring serialization to JSON. The interface is that of a stack, in which values are pushed and then various operations are used to turn terminal types into more complex ones like objects and arrays. When used, the Input will provide the top of the stack to any downstream consumer (such as Interpreter::set_input()).

Interpreter

Interface for the Rego interpreter.

Node

Interface for Rego Node objects.

Output

Interface for the Rego output.

__fsid_t

Enums

BundleFormat

Format types for serializing bundles.

LogLevel

NodeKind

Enumeration of different kinds of Rego Nodes that can be returned as output from a query.

NodeValue

Represents the value of a Rego Node.

Constants

INT8_MAX

INT8_MIN

INT16_MAX

INT16_MIN

INT32_MAX

INT32_MIN

INTPTR_MAX

INTPTR_MIN

INT_FAST8_MAX

INT_FAST8_MIN

INT_FAST16_MAX

INT_FAST16_MIN

INT_FAST32_MAX

INT_FAST32_MIN

INT_LEAST8_MAX

INT_LEAST8_MIN

INT_LEAST16_MAX

INT_LEAST16_MIN

INT_LEAST32_MAX

INT_LEAST32_MIN

PTRDIFF_MAX

PTRDIFF_MIN

REGOCPP_BUILD_DATE

REGOCPP_BUILD_NAME

REGOCPP_BUILD_TOOLCHAIN

REGOCPP_GIT_HASH

REGOCPP_OPA_VERSION

REGOCPP_PLATFORM

REGOCPP_VERSION

REGOCPP_VERSION_MAJOR

REGOCPP_VERSION_MINOR

REGOCPP_VERSION_REVISION

REGO_BUILD_INFO

REGO_ERROR

REGO_ERROR_BUFFER_TOO_SMALL

REGO_ERROR_INPUT_MISSING_ARGUMENTS

REGO_ERROR_INPUT_NULL

REGO_ERROR_INPUT_OBJECT_ITEM

REGO_ERROR_INVALID_LOG_LEVEL

REGO_LOG_LEVEL_DEBUG

REGO_LOG_LEVEL_ERROR

REGO_LOG_LEVEL_INFO

REGO_LOG_LEVEL_NONE

REGO_LOG_LEVEL_OUTPUT

REGO_LOG_LEVEL_TRACE

REGO_LOG_LEVEL_UNSUPPORTED

REGO_LOG_LEVEL_WARN

REGO_NODE_ARRAY

REGO_NODE_BINDING

REGO_NODE_BINDINGS

REGO_NODE_ERROR

REGO_NODE_ERROR_AST

REGO_NODE_ERROR_CODE

REGO_NODE_ERROR_MESSAGE

REGO_NODE_ERROR_SEQ

REGO_NODE_FALSE

REGO_NODE_FLOAT

REGO_NODE_INT

REGO_NODE_INTERNAL

REGO_NODE_NULL

REGO_NODE_OBJECT

REGO_NODE_OBJECT_ITEM

REGO_NODE_RESULT

REGO_NODE_RESULTS

REGO_NODE_SCALAR

REGO_NODE_SET

REGO_NODE_STRING

REGO_NODE_TERM

REGO_NODE_TERMS

REGO_NODE_TRUE

REGO_NODE_UNDEFINED

REGO_NODE_VAR

REGO_OK

SIG_ATOMIC_MAX

SIG_ATOMIC_MIN

SIZE_MAX

UINT8_MAX

UINT16_MAX

UINT32_MAX

UINTPTR_MAX

UINT_FAST8_MAX

UINT_FAST16_MAX

UINT_FAST32_MAX

UINT_LEAST8_MAX

UINT_LEAST16_MAX

UINT_LEAST32_MAX

WINT_MAX

WINT_MIN

_ATFILE_SOURCE

_BITS_STDINT_INTN_H

_BITS_STDINT_LEAST_H

_BITS_STDINT_UINTN_H

_BITS_TIME64_H

_BITS_TYPESIZES_H

_BITS_TYPES_H

_BITS_WCHAR_H

_DEFAULT_SOURCE

_FEATURES_H

_POSIX_C_SOURCE

_POSIX_SOURCE

_STDC_PREDEF_H

_STDINT_H

_SYS_CDEFS_H

__FD_SETSIZE

__GLIBC_MINOR__

__GLIBC_USE_C2X_STRTOL

__GLIBC_USE_DEPRECATED_GETS

__GLIBC_USE_DEPRECATED_SCANF

__GLIBC_USE_IEC_60559_BFP_EXT

__GLIBC_USE_IEC_60559_BFP_EXT_C2X

__GLIBC_USE_IEC_60559_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT

__GLIBC_USE_IEC_60559_FUNCS_EXT_C2X

__GLIBC_USE_IEC_60559_TYPES_EXT

__GLIBC_USE_ISOC2X

__GLIBC_USE_LIB_EXT2

__GLIBC__

__GNU_LIBRARY__

__HAVE_GENERIC_SELECTION

__INO_T_MATCHES_INO64_T

__KERNEL_OLD_TIMEVAL_MATCHES_TIMEVAL64

__LDOUBLE_REDIRECTS_TO_FLOAT128_ABI

__OFF_T_MATCHES_OFF64_T

__RLIM_T_MATCHES_RLIM64_T

__STATFS_MATCHES_STATFS64

__STDC_IEC_559_COMPLEX__

__STDC_IEC_559__

__STDC_IEC_60559_BFP__

__STDC_IEC_60559_COMPLEX__

__STDC_ISO_10646__

__SYSCALL_WORDSIZE

__TIMESIZE

__USE_ATFILE

__USE_FORTIFY_LEVEL

__USE_ISOC11

__USE_ISOC95

__USE_ISOC99

__USE_MISC

__USE_POSIX

__USE_POSIX2

__USE_POSIX199309

__USE_POSIX199506

__USE_POSIX_IMPLICITLY

__USE_XOPEN2K

__USE_XOPEN2K8

__WORDSIZE

__WORDSIZE_TIME64_COMPAT32

__glibc_c99_flexarr_available

Functions

build_info

Returns the build information as a string.

log_level_from_string

regoAddDataJSON⚠

@brief Adds a base document from the specified string. @details The string should contain a single JSON object. The object will be parsed and merged with the interpreter’s base document. @par If an error code is returned, more error information can be obtained by calling ::regoError. @param rego The interpreter. @param contents The contents of the JSON object. @return REGO_OK if successful, REGO_ERROR otherwise.

regoAddDataJSONFile⚠

@brief Adds a base document from the file at the specified path. @details The file should contain a single JSON object. The object will be parsed and merged with the interpreter’s base document. @par If an error code is returned, more error information can be obtained by calling ::regoError. @param rego The interpreter. @param path The path to the JSON file. @return REGO_OK if successful, REGO_ERROR otherwise.

regoAddEntrypoint⚠

@brief Adds an entrypoint to the interpreter. @details By using this method, an entrypoint can be added to the list of those with compiled execution plans in a bundle. @param rego The interpreter @param entrypoint A valid Rego entrypoint of the form path/to/rule @return REGO_OK if successful, REGO_ERROR if there were any issues adding the entrypoint.

regoAddModule⚠

@brief Adds a module (e.g. virtual document) from the specified string. @details If an error code is returned, more error information can be obtained by calling ::regoError. @param rego The interpreter. @param name The name of the module. @param contents The contents of the module. @return REGO_OK if successful, REGO_ERROR otherwise.

regoAddModuleFile⚠

@brief Adds a module (e.g. virtual document) from the file at the specified path. @details If an error code is returned, more error information can be obtained by calling ::regoError. @param rego The interpreter. @param path The path to the policy file. @return REGO_OK if successful, REGO_ERROR otherwise.

regoBuild⚠

@brief Builds a compiled Rego bundle. @details The bundle will include any base and virtual documents added to the interpreter, as well as a query set with ::regoSetQuery (if present) and any entrypoints added with ::regoAddEntrypoint (if any). There must be at least one entrypoint or a query specified for the build to be successful. The resulting handle can be used with ::regoBundleQuery and/or saved to the disk for later re-use. @return a compiled Rego bundle

regoBuildInfo⚠

@brief Populate a buffer with a string of the form “VERSION (BUILD_NAME, BUILD_DATE) BUILD_TOOLCHAIN on PLATFORM” @details The buffer must be large enough to hold the value. The size of the buffer can be determined by calling ::regoBuildInfoSize. @param buffer The buffer to populate. @param size The size of the buffer. @return REGO_OK if successful, REGO_ERROR_BUFFER_TOO_SMALL otherwise.

regoBuildInfoSize⚠

@brief Returns the number of bytes needed to store a 0-terminated string representing the build info. @details The value returned by this function can be used to allocate a buffer to pass to ::regoBuildInfo. @return The number of bytes needed to store the build info.

regoBundleLoad⚠

@brief Loads a compiled Rego bundle from the specified directory. @details The directory must contain a valid Rego bundle, which consists of a plan.json file, a data.json file, and zero or more .rego files which were used to create the bundle. @param rego The interpreter @param dir The path to the directory containing the bundle. @return The loaded bundle, or NULL if there was an error.

regoBundleLoadBinary⚠

@brief Loads a compiled Rego bundle from the specified binary file. @details The file must contain a valid Rego bundle in Rego Bundle Binary format. For more information on the format, see the specification. @param rego The interpreter @param path The path to the binary file containing the bundle. @return The loaded bundle, or NULL if there was an error.

regoBundleNode⚠

@brief Returns a node containing the base document of the bundle. @warning If the node was loaded from a binary file, this node will be missing. Future versions of the API may provide a way to reconstruct the bundle AST from the binary file. @param bundle The bundle. @return A node containing the base document of the bundle.

regoBundleOk⚠

@brief Returns whether the bundle is ok. @details If the bundle was built or loaded successfully, this function will return true. Otherwise, it will return false, indicating that there was an error during bundle creation. This error can be retrieved using regoError on the interpreter used to create or load the bundle. @param bundle The bundle. @return Whether the bundle is ok.

regoBundleQuery⚠

@brief Performs a query against the specified bundle. @details The output of the query will be returned as a regoOutput object. The query executed will be the one set with ::regoSetQuery when the bundle was built. If no query was set, the output will be an error sequence. @note The caller is responsible for freeing the output object with ::regoFreeOutput. @param rego The interpreter. @param bundle The bundle to query. @return The output of the query.

regoBundleQueryEntrypoint⚠

@brief Performs a query against the specified bundle at the specified entrypoint. @details The output of the query will be returned as a regoOutput object. The entrypoint must be one of those added with ::regoAddEntrypoint when the bundle was built. If the entrypoint does not exist, the output will be an error sequence. @note The caller is responsible for freeing the output object with ::regoFreeOutput. @param rego The interpreter. @param bundle The bundle to query. @param endpoint The entrypoint to query. @return The output of the query.

regoBundleSave⚠

@brief Saves a compiled Rego bundle to the specified directory. @details The directory will be created if it does not exist. The bundle will be saved in the standard Rego bundle format, which consists of a plan.json file, a data.json file, and zero or more .rego files. @param rego The interpreter @param dir The path to the directory where the bundle will be saved. @param bundle The bundle to save. @return REGO_OK if successful, REGO_ERROR otherwise.

regoBundleSaveBinary⚠

@brief Saves a compiled Rego bundle to the specified binary file. @details The bundle will be saved in Rego Bundle Binary format. For more information on the format, see the specification. @param rego The interpreter @param path The path to the binary file where the bundle will be saved. @param bundle The bundle to save. @return REGO_OK if successful, REGO_ERROR otherwise.

regoError⚠

@brief Populate a buffer with the most recent error message. @details If an error code is returned from an interface function, more error information can usually be obtained by calling this function. @par The buffer must be large enough to hold the value. The size of the buffer can be determined by calling ::regoErrorSize. @param rego The interpreter. @param buffer The buffer to populate. @param size The size of the buffer. @return REGO_OK if successful, REGO_ERROR_BUFFER_TOO_SMALL otherwise.

regoErrorSize⚠

@brief Returns the number of bytes needed to store a 0-terminated string representing the most recent error message. @details The value returned by this function can be used to allocate a buffer to pass to ::regoError. @return The number of bytes needed to store the error message.

regoFree⚠

@brief Frees a Rego interpreter. @note This pointer must have been allocated with ::regoNew. @param rego The interpreter to free.

regoFreeBundle⚠

@brief Frees a Rego bundle. @note This pointer must have been allocated with ::regoBuild, ::regoBundleLoad, or ::regoBundleLoadBinary. @param bundle The bundle to free.

regoFreeInput⚠

@brief Frees a Rego Input object. @note This pointer must have been allocated with ::regoNewInput. @param input The input object to free.

regoFreeOutput⚠

@brief Frees a Rego output. @note This pointer must have been allocated with regoQuery. @param output The output to free.

regoGetDebugEnabled⚠

regoGetLogLevel⚠

@brief Gets the level of logging on the interpreter. @return One of the following values: REGO_LOG_LEVEL_NONE, REGO_LOG_LEVEL_ERROR, REGO_LOG_LEVEL_OUTPUT, REGO_LOG_LEVEL_WARN, REGO_LOG_LEVEL_INFO, REGO_LOG_LEVEL_DEBUG, REGO_LOG_LEVEL_TRACE.

regoGetStrictBuiltInErrors⚠

@brief Gets whether strict built-in errors are enabled. @param rego The interpreter. @return Whether strict built-in errors are enabled.

regoGetWellFormedChecksEnabled⚠

@brief Gets whether well-formedness checks are enabled. @param rego The interpreter. @return Whether well-formedness checks are enabled.

regoInputArray⚠

@brief Uses the top ‘count’ values on the stack to create an array and pushes it back onto the stack. @param input The input object. @param count The number of values to use. @return REGO_OK if successful, REGO_ERROR or REGO_INPUT_MISSING_ARGUMENTS otherwise.

regoInputBoolean⚠

@brief Adds a boolean value to the input object. @param input The input object. @param value The boolean value. @return REGO_OK if successful, REGO_ERROR otherwise.

regoInputFloat⚠

@brief Adds a floating point value to the input object. @param input The input object. @param value The floating point value. @return REGO_OK if successful, REGO_ERROR otherwise.

regoInputInt⚠

@brief Adds an integer value to the input object. @param input The input object. @param value The integer value. @return REGO_OK if successful, REGO_ERROR otherwise.

regoInputNode⚠

@brief Returns the top of the Input stack. @param input The input object. @return The top of the input stack as a regoNode. If the input is invalid or empty, this function will return NULL.

regoInputNull⚠

@brief Adds a null value to the input object. @param input The input object. @return REGO_OK if successful, REGO_ERROR otherwise.

regoInputObject⚠

@brief Uses the top ‘count’ values on the stack to create an object and pushes it back onto the stack. @note The top ‘count’ values must be object items. @param input The input object. @param count The number of object items to use. @return REGO_OK if successful, REGO_ERROR, REGO_INPUT_MISSING_ARGUMENTS, or REGO_INPUT_OBJECT_ITEM otherwise.

regoInputObjectItem⚠

@brief Uses the top two values on the stack to create an object item and pushes it back onto the stack. @details The top value on the stack will be used as the value, and the second value will be used as the key. @param input The input object. @return REGO_OK if successful, REGO_ERROR or REGO_INPUT_MISSING_ARGUMENTS otherwise.

regoInputSet⚠

@brief Uses the top ‘count’ values on the stack to create a set and pushes it back onto the stack. @note As part of this operation, duplicate values will be removed and the values will be sorted. @param input The input object. @param count The number of values to use. @return REGO_OK if successful, REGO_ERROR or REGO_INPUT_MISSING_ARGUMENTS otherwise.

regoInputSize⚠

@brief Returns the number of items on the Input stack. @param input The input object. @return The number of items on the input stack.

regoInputString⚠

@brief Adds a string value to the input object. @param input The input object. @param value The string value. @return REGO_OK if successful, REGO_ERROR otherwise.

regoInputValidate⚠

@brief Validates the input object. @details This function should be called after building the input object and before using it with ::regoSetInput. This function will ensure that the input object is valid (e.g. that there is exactly one value on the stack and there were no errors during construction). @param input The input object. @return REGO_OK if the input is valid, REGO_ERROR_INPUT_* otherwise.

regoIsAvailableBuiltIn⚠

@brief Returns whether the specified name corresponds to an available built-in in the interpreter. @param rego The interpreter. @param name The name of the built-in. @return Whether the built-in exists.

regoLogLevelFromString⚠

@brief Gets a log level constant value from a string (case-invariant). @details String must be one of None, Error, Output, Warn, Info, Debug or Trace. An unrecognized string will result in a value of REGO_LOG_LEVEL_UNSUPPORTED. @param level The log level string. @return One of the following values: REGO_LOG_LEVEL_NONE, REGO_LOG_LEVEL_ERROR, REGO_LOG_LEVEL_OUTPUT, REGO_LOG_LEVEL_WARN, REGO_LOG_LEVEL_INFO, REGO_LOG_LEVEL_DEBUG, REGO_LOG_LEVEL_TRACE, or REGO_LOG_LEVEL_UNSUPPORTED

regoNew⚠

@brief Allocates and initializes a new Rego interpreter. @note The caller is responsible for freeing the interpreter with ::regoFree. @return A pointer to the new interpreter.

regoNewInput⚠

@brief Allocates and initializes a new Rego Input object. @details Rego Input objects can be used to build input documents programmatically. An input is built up by making a series of calls to the API to add values to the stack. Some operations use the values on the stack to create composite values (e.g. objects, arrays, sets) which are then pushed back onto the stack. Once the input is built, it can be validated and then set on the interpreter with regoSetInput. While you can test the result of every operation, you can continue to use a broken Input object and the error will be returned when you call regoInputValidate.

regoNodeGet⚠

@brief Returns the child node at the specified index. @param node The node. @param index The index of the child. @return The child node.

regoNodeJSON⚠

@brief Populate a buffer with the JSON representation of the node. @details The buffer must be large enough to hold the value. The size of the buffer can be determined by calling ::regoNodeJSONSize. @param node The node. @param buffer The buffer to populate. @param size The size of the buffer. @return REGO_OK if successful, REGO_ERROR_BUFFER_TOO_SMALL otherwise.

regoNodeJSONSize⚠

@brief Returns the number of bytes needed to store a 0-terminated string representing the JSON representation of the node. @details The value returned by this function can be used to allocate a buffer to pass to ::regoNodeJSON. @param node The node. @return The number of bytes needed to store the JSON representation.

regoNodeSize⚠

@brief Returns the number of children of the node. @param node The node. @return The number of children.

regoNodeType⚠

@brief Returns an enumeration value indicating the node’s type. @details This type will be one of the following values:

regoNodeTypeName⚠

@brief Returns the name of the node type as a human-readable string. @details This function supports arbitrary nodes (i.e. it will always produce a value) including internal nodes which appear in error messages. The buffer must be large enough to hold the value. The size of the buffer can be determined by calling ::regoNodeTypeNameSize. @param node The node. @param buffer The buffer to populate. @param size The size of the buffer. @return REGO_OK if successful, REGO_ERROR_BUFFER_TOO_SMALL otherwise.

regoNodeTypeNameSize⚠

@brief Returns the number of bytes needed to store a 0-terminated string representing the name of the node type. @details The value returned by this function can be used to allocate a buffer to pass to ::regoNodeTypeName. @param node The node. @return The number of bytes needed to store the node type name.

regoNodeValue⚠

@brief Populate a buffer with the node value. @details The buffer must be large enough to hold the value. The size of the buffer can be determined by calling regoNodeValueSize. @param node The node. @param buffer The buffer to populate. @param size The size of the buffer. @return REGO_OK if successful, REGO_ERROR_BUFFER_TOO_SMALL otherwise.

regoNodeValueSize⚠

@brief Returns the number of bytes needed to store a 0-terminated string representing the text value of the node. @details The value returned by this function can be used to allocate a buffer to pass to ::regoNodeValue. @param node The node. @return The number of bytes needed to store the text value.

regoOutputBinding⚠

@brief Returns the bound value for a given variable name at the first index. @details If the variable is not bound, then this function will return NULL. @param output The output. @param name The variable name. @return The bound value (or NULL if the variable was not bound)

regoOutputBindingAtIndex⚠

@brief Returns the bound value for a given variable name. @details If the variable is not bound, then this function will return NULL. @param output The output. @param index The result index. @param name The variable name. @return The bound value (or NULL if the variable was not bound)

regoOutputExpressions⚠

@brief Returns a node containing a list of terms resulting from the query at the default index. @param output The output. @return The output node.

regoOutputExpressionsAtIndex⚠

@brief Returns a node containing a list of terms resulting from the query at the specified index. @param output The output. @param index The result index. @return The output node.

regoOutputJSON⚠

@brief Populate a buffer with the output represented as a human-readable string. @details The buffer must be large enough to hold the value. The size of the buffer can be determined by calling ::regoOutputJSONSize. @param output The output. @param buffer The buffer to populate. @param size The size of the buffer. @return REGO_OK if successful, REGO_ERROR_BUFFER_TOO_SMALL otherwise.

regoOutputJSONSize⚠

@brief Returns the number of bytes needed to store a 0-terminated string representing the output as a human-readable string. @details The value returned by this function can be used to allocate a buffer to pass to ::regoOutputJSON. @param output The output. @return The number of bytes needed to store the output string.

regoOutputNode⚠

@brief Returns the node containing the output of the query. @details This will either be a node which contains one or more results, or an error sequence. @param output The output. @return The output node.

regoOutputOk⚠

@brief Returns whether the output is ok. @details If the output resulted in a valid query result, then this function will return true. Otherwise, it will return false, indicating that the output contains an error sequence. @param output The output. @return Whether the output is ok.

regoOutputSize⚠

@brief Returns the number of results in the output. @details Each query can potentially generate multiple results. @return The number of results in the output

regoQuery⚠

@brief Performs a query against the current base and virtual documents. @details The query expression should be a Rego query. The output of the query will be returned as a regoOutput object. @note The caller is responsible for freeing the output object with ::regoFreeOutput. @param rego The interpreter. @param query_expr The query expression. @return The output of the query.

regoSetDebugEnabled⚠

@brief Sets the debug mode of the interpreter. @details When debug mode is enabled, the interpreter will output intermediary ASTs after each compiler pass to the debug directory and output pass information to stdout. This is mostly useful for creating reports for compiler issues, but can also be of use in understanding why a policy is invalid or is behaving unexpectedly. @param rego The interpreter. @param enabled Whether debug mode should be enabled.

regoSetDebugPath⚠

@brief Sets the path to the debug directory. @details If set, then (when in debug mode) the interpreter will output intermediary ASTs after each compiler pass to the debug directory. If the directory does not exist, it will be created. @par If an error code is returned, more error information can be obtained by calling ::regoError. @param rego The interpreter. @param path The path to the debug directory. @return REGO_OK if successful, REGO_ERROR otherwise.

regoSetDefaultLogLevel⚠

@brief Sets the default level of logging. @details This setting controls the amount of logging that will be output to stdout. The default level is REGO_LOG_LEVEL_OUTPUT. New interpreters will be set to this level of logging unless overridden by ::regoSetLogLevel. @param level One of the following values: REGO_LOG_LEVEL_NONE, REGO_LOG_LEVEL_ERROR, REGO_LOG_LEVEL_OUTPUT, REGO_LOG_LEVEL_WARN, REGO_LOG_LEVEL_INFO, REGO_LOG_LEVEL_DEBUG, REGO_LOG_LEVEL_TRACE. @return REGO_OK if successful, REGO_ERROR_INVALID_LOG_LEVEL otherwise.

regoSetInput⚠

@brief Sets the current input document from the specified Input object. @details If an error code is returned, more error information can be obtained by calling ::regoError. @param rego The interpreter. @param input The input object. @return REGO_OK if successful, REGO_ERROR otherwise.

regoSetInputJSONFile⚠

@brief Sets the current input document from the file at the specified path. @details The file should contain a single JSON value. The value will be parsed and set as the interpreter’s input document. @par If an error code is returned, more error information can be obtained by calling ::regoError. @param rego The interpreter. @param path The path to the JSON file. @return REGO_OK if successful, REGO_ERROR otherwise.

regoSetInputTerm⚠

@brief Sets the current input document from the specified string. @details The string should contain a single Rego data term. The value will be parsed and set as the interpreter’s input document. @par If an error code is returned, more error information can be obtained by calling ::regoError. @param rego The interpreter. @param contents The contents of the Rego data term. @return REGO_OK if successful, REGO_ERROR otherwise.

regoSetLogLevel⚠

@brief Sets the level of logging on the interpreter. @details This setting controls the amount of logging that will be output to stdout. The default level is REGO_LOG_LEVEL_OUTPUT, but can be changed using ::regoSetDefaultLogLevel. @param rego The interpreter whose log level is to be set. @param level One of the following values: REGO_LOG_LEVEL_NONE, REGO_LOG_LEVEL_ERROR, REGO_LOG_LEVEL_OUTPUT, REGO_LOG_LEVEL_WARN, REGO_LOG_LEVEL_INFO, REGO_LOG_LEVEL_DEBUG, REGO_LOG_LEVEL_TRACE. @return REGO_OK if successful, REGO_ERROR_INVALID_LOG_LEVEL otherwise.

regoSetQuery⚠

@brief Sets the query for the interpreter. @details By using this method, a query can be set such that it is included as a default entrypoint in a bundle. @param rego The interpreter @param query_expr A Rego query expression @return REGO_OK if successful, REGO_ERROR if there was a problem parsing the query

regoSetStrictBuiltInErrors⚠

@brief Sets whether the built-ins should throw errors. @details When strict built-in errors are enabled, built-in functions will throw errors when they encounter invalid input. When disabled, built-in functions will return undefined when they encounter invalid input. @param rego The interpreter. @param enabled Whether strict built-in errors should be enabled.

regoSetWellFormedChecksEnabled⚠

@brief Sets whether to perform well-formedness checks after each compiler pass. @details The interpreter has a set of well-formedness definitions which indicate the expected form of the AST before and after each compiler pass. This setting determines whether the interpreter will perform these intermediary checks. @param rego The interpreter. @param enabled Whether well-formedness checks should be enabled.

regoVersion⚠

@brief Populate a buffer with a string representing the semantic version of the library. @details The buffer must be large enough to hold the value. The size of the buffer can be determined by calling ::regoVersionSize. @param buffer The buffer to populate. @param size The size of the buffer. @return REGO_OK if successful, REGO_ERROR_BUFFER_TOO_SMALL otherwise.

regoVersionSize⚠

@brief Returns the number of bytes needed to store a 0-terminated string representing version. @details The value returned by this function can be used to allocate a buffer to pass to ::regoVersion. @return The number of bytes needed to store the version string.

Type Aliases

__blkcnt64_t

__blkcnt_t

__blksize_t

__caddr_t

__clock_t

__clockid_t

__daddr_t

__dev_t

__fsblkcnt64_t

__fsblkcnt_t

__fsfilcnt64_t

__fsfilcnt_t

__fsword_t

__gid_t

__id_t

__ino64_t

__ino_t

__int8_t

__int16_t

__int32_t

__int64_t

__int_least8_t

__int_least16_t

__int_least32_t

__int_least64_t

__intmax_t

__intptr_t

__key_t

__loff_t

__mode_t

__nlink_t

__off64_t

__off_t

__pid_t

__quad_t

__rlim64_t

__rlim_t

__sig_atomic_t

__socklen_t

__ssize_t

__suseconds64_t

__suseconds_t

__syscall_slong_t

__syscall_ulong_t

__time_t

__timer_t

__u_char

__u_int

__u_long

__u_quad_t

__u_short

__uid_t

__uint8_t

__uint16_t

__uint32_t

__uint64_t

__uint_least8_t

__uint_least16_t

__uint_least32_t

__uint_least64_t

__uintmax_t

__useconds_t

int_fast8_t

int_fast16_t

int_fast32_t

int_fast64_t

int_least8_t

int_least16_t

int_least32_t

int_least64_t

intmax_t

regoBoolean

@brief Boolean type

regoBundle

@brief Opaque bundle type

regoEnum

@brief Enum type

regoInput

@brief Opaque input type

regoInt

@brief Integer type

regoInterpreter

@brief Opaque interpreter type

regoNode

@brief Opaque node type

regoOutput

@brief Opaque output type

regoSize

@brief Size type

regoType

@brief Node Type type

uint_fast8_t

uint_fast16_t

uint_fast32_t

uint_fast64_t

uint_least8_t

uint_least16_t

uint_least32_t

uint_least64_t

uintmax_t